Show provider availabilities
Currently, prospective clients have to click each provider's page individually to determine who, if anyone, has availabilities.

This change shows a line of text saying either "❎ No availabilities at this time" or a check (✅) followed by the provider's availabilities. (As specified on the individual provider pages in the `_psychotherapy` directory.)

This change also adds an `any_availability` variable to each provider; I filled in the values as best I could infer based on the `availability` fields.

User story:
As a: prospective patient
I want: to know which providers have availabilities without opening 20 new tabs
So that: I can quickly determine if it's worth anyone's time to send an email